refactor : Optimize file and folder directories
