+++ lib/screen/tabs/map_tab/search_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/container.dart';
 import 'package:flutter/src/widgets/framework.dart';
-import 'package:coffeeconti/associated_search.dart';
+import 'package:coffeeconti/components/button/associated_search_button.dart';
 
 const interests = [
   'Apple',
