+++ lib/widgets/keyword_widget.dart
@@ -1,88 +1,52 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:provider/provider.dart';
 
 import '../components/button/keyword_button.dart';
-import '../data/models/keyword_model.dart';
+import '../data/provider/keyword_provider.dart';
+import '../popup/show_category_sheet.dart';
 
-class KeywordWidget extends StatefulWidget {
-  final List<KeywordData> keywordDatas;
-  const KeywordWidget({
-    super.key,
-    required this.keywordDatas,
-  });
+class KeywordWidget extends StatelessWidget {
+  const KeywordWidget({super.key});
 
-  @override
-  State<KeywordWidget> createState() => _KeywordWidgetState();
-}
-
-class _KeywordWidgetState extends State<KeywordWidget> {
-  final ScrollController _scrollController = ScrollController();
-
-  bool _showTitle = false;
-
-  void _onScroll() {
-    if (_scrollController.offset > 100) {
-      if (_showTitle) return;
-      setState(() {
-        _showTitle = true;
-      });
-    } else {
-      setState(() {
-        _showTitle = false;
-      });
+  void _tapCategory(BuildContext context, int index) {
+    if (index == 0) {
+      showModalBottomSheet(
+        context: context,
+        isScrollControlled: true,
+        builder: (context) {
+          return ShowCategorySheet();
+        },
+      );
     }
   }
 
   @override
-  void initState() {
-    super.initState();
-    _scrollController.addListener(() {
-      _onScroll();
-    });
-  }
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return Expanded(
-      child: Padding(
-        padding: EdgeInsets.symmetric(
-          vertical: 20.0.h,
-          horizontal: 20.0.w,
-        ),
-        child: Container(
-          //color: Colors.pink,
-          child: Scrollbar(
-            controller: _scrollController,
-            child: SingleChildScrollView(
-              controller: _scrollController,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Wrap(
-                    alignment: WrapAlignment.start,
-                    runSpacing: 18,
-                    spacing: 10,
-                    children: [
-                      for (var keywordData in widget.keywordDatas)
-                        KeywordButton(
-                          keywordName: keywordData.name,
-                          keywordId: keywordData.id,
-                        )
-                    ],
-                  )
-                ],
+    return Consumer<KeywordsProvider>(
+      builder: (context, provider, child) {
+        return Padding(
+          padding: EdgeInsets.symmetric(vertical: 10.0.h),
+          child: SizedBox(
+            height: 40.0.h,
+            child: ListView.builder(
+              shrinkWrap: true,
+              itemCount: provider.selectedKeywordsName.length + 1,
+              scrollDirection: Axis.horizontal,
+              itemBuilder: (context, index) => KeywordButton(
+                name: provider.selectedCategoryId == null
+                    ? '카테고리'
+                    : index == 0
+                        ? '${provider.categoryNames[provider.selectedCategoryId!]}'
+                        : '${provider.selectedKeywordsName[index - 1]}',
+                onTap: () {
+                  _tapCategory(context, index);
+                },
               ),
             ),
           ),
-        ),
-      ),
+        );
+      },
     );
   }
 }
