--- lib/widgets/tabs/profile.dart
@@ -1,17 +1,0 @@
-
-// // controller.addOverlayAll{ NMarker(
-// //                 id: 'test',
-// //                 position:
-// //                     const NLatLng(37.506932467450326, 127.05578661133796))
-
-// //                   , NMarker(
-// //                 id: 'test1',
-// //                 position:
-// //                     const NLatLng(37.606932467450326, 127.05578661133796))}
-// mapController.addOverlayAll({
-//         NMarker(id: 'test', position: NLatLng(37.612941, 127.039381)),
-//         NMarker(
-//                id: 'test1',
-//                  position:
-//                     const NLatLng(37.606932467450326, 127.05578661133796))
-//       }
